big commit, HomePage.js, favoritePage, router, moved character to CharacterPage,  NOT DONE YET
